More cleanup; create energy_cap group constraint test file
