+++ calliope/test/common/test_model/model_config_group/base_model.yaml
@@ -1,0 +1,92 @@
+import:
+    - scenarios_add_techs.yaml
+    - scenarios_energy_cap.yaml
+
+model:
+    name: Group share constraint test model
+    timeseries_data_path: '../timeseries_data'
+
+run:
+    mode: plan
+    solver: cplex
+    objective_options.cost_class: {monetary: 1}
+
+tech_groups:
+    cheap_supply:
+        essentials:
+            parent: supply
+        constraints:
+            lifetime: 25
+            energy_cap_max: .inf
+            resource: file=supply_simple.csv
+        costs:
+            monetary:
+                interest_rate: 0.1
+                energy_cap: 10
+    expensive_supply:
+        essentials:
+            parent: supply
+        constraints:
+            lifetime: 25
+            energy_cap_max: .inf
+        costs:
+            monetary:
+                interest_rate: 0.1
+                energy_cap: 20
+    normal_supply:
+        essentials:
+            parent: supply
+        constraints:
+            lifetime: 25
+            energy_cap_max: .inf
+        costs:
+            monetary:
+                interest_rate: 0.1
+                energy_cap: 15
+
+techs:
+    cheap_elec_supply:
+        essentials:
+            name: Cheap electricity supply
+            parent: cheap_supply
+            carrier: electricity
+
+    expensive_elec_supply:
+        essentials:
+            name: Expensive electricity supply
+            parent: expensive_supply
+            carrier: electricity
+
+    normal_elec_supply:
+        essentials:
+            name: Normal electricity supply
+            parent: normal_supply
+            carrier: electricity
+
+    electricity_demand:
+        essentials:
+            name: Electricity demand
+            carrier: electricity
+            parent: demand
+        constraints:
+            resource: file=demand_simple.csv
+
+    cheap_elec_transmission:
+        essentials:
+            name: 'Power transmission'
+            parent: transmission
+            carrier: electricity
+        constraints:
+            energy_cap_max: inf
+            energy_eff: 0.98
+            lifetime: 25
+        costs:
+            monetary:
+                interest_rate: 0.1
+                energy_cap: 5
+
+locations:
+    0,1.techs: {cheap_elec_supply, expensive_elec_supply, normal_elec_supply, electricity_demand}
+
+links:
+    0,1.techs: {cheap_elec_transmission}
