+++ calliope/test/common/test_model/model_config_group/scenarios_add_techs.yaml
@@ -1,0 +1,145 @@
+overrides:
+    add_cheap_heat_supply:
+        techs.cheap_heat_supply.essentials:
+            name: Cheap heating
+            parent: cheap_supply
+            carrier: heat
+        locations.1.techs: {cheap_heat_supply}
+
+    add_normal_heat_supply:
+        techs.normal_heat_supply.essentials:
+            name: Normal heating
+            parent: normal_supply
+            carrier: heat
+        locations.1.techs: {normal_heat_supply}
+
+    add_expensive_heat_supply:
+        techs.expensive_heat_supply.essentials:
+            name: Expensive heating
+            parent: expensive_supply
+            carrier: heat
+        locations.1.techs: {expensive_heat_supply}
+
+    add_cheap_cooling_supply:
+        techs.cheap_cool_supply.essentials:
+            name: Cheap cooling
+            parent: cheap_supply
+            carrier: cool
+        locations.1.techs: {cheap_cool_supply}
+
+    add_normal_cooling_supply:
+        techs.normal_cool_supply.essentials:
+            name: Normal cooling
+            parent: normal_supply
+            carrier: cool
+        locations.1.techs: {normal_cool_supply}
+
+    add_expensive_cooling_supply:
+        techs.expensive_cool_supply.essentials:
+            name: Expensive cooling
+            parent: expensive_supply
+            carrier: cool
+        locations.1.techs: {expensive_cool_supply}
+
+    add_elec_to_heat_conversion:
+        techs.elec_to_heat:
+            essentials:
+                name: 'Electric heating'
+                parent: conversion
+                carrier_in: electricity
+                carrier_out: heat
+            constraints:
+                energy_cap_max: .inf
+                energy_eff: 0.95
+                lifetime: 25
+            costs:
+                monetary:
+                    interest_rate: 0.1
+                    om_con: 0.001
+        locations.1.techs: {elec_to_heat}
+
+    add_elec_to_heat_cooling_linked_conversion_plus:
+        techs.elec_to_heat_cool_linked:
+            essentials:
+                name: 'Linked electric heating and cooling'
+                parent: conversion_plus
+                carrier_in: electricity
+                carrier_out: heat
+                carrier_out_2: cool
+                primary_carrier_out: heat
+            constraints:
+                energy_cap_max: .inf
+                energy_eff: 0.95
+                lifetime: 25
+                carrier_ratios.carrier_out_2.cool: 0.5
+            costs:
+                monetary:
+                    interest_rate: 0.1
+                    om_con: 0.001
+        locations.1.techs: {elec_to_heat_cool_linked}
+
+    add_elec_to_heat_cooling_unlinked_conversion_plus:
+        techs.elec_to_heat_cool_unlinked:
+            essentials:
+                name: 'Unlinked electric heating and cooling'
+                parent: conversion_plus
+                carrier_in: electricity
+                carrier_out: [heat, cool]
+                primary_carrier_out: heat
+            constraints:
+                energy_cap_max: .inf
+                energy_eff: 0.95
+                lifetime: 25
+                carrier_ratios.carrier_out: {heat: 1, cool: 0.5}
+            costs:
+                monetary:
+                    interest_rate: 0.1
+                    om_con: 0.001
+        locations.1.techs: {elec_to_heat_cool_unlinked}
+
+    add_heat_demand:
+        techs.heat_demand:
+            essentials:
+                name: Heat demand
+                carrier: heat
+                parent: demand
+            constraints:
+                resource: -2.5
+        locations.1.techs: {heat_demand}
+
+    add_cooling_demand:
+        techs.cool_demand:
+            essentials:
+                name: Cooling demand
+                carrier: cool
+                parent: demand
+            constraints:
+                resource: -1
+        locations.1.techs: {cool_demand}
+
+    add_expensive_elec_transmission:
+        techs.expensive_elec_transmission:
+            essentials:
+                name: Expensive power transmission
+                parent: transmission
+                carrier: electricity
+            constraints:
+                energy_cap_max: inf
+                energy_eff: 0.98
+                lifetime: 25
+            costs:
+                monetary:
+                    interest_rate: 0.1
+                    energy_cap: 10
+        links.0,1.techs: {expensive_elec_transmission}
+
+    add_elec_storage:
+        techs.elec_storage:
+            essentials:
+                name: Electricity storage
+                carrier: electricity
+                parent: storage
+            constraints:
+                storage_loss: 0.001
+                energy_eff: 0.99
+        locations.1.techs: {elec_storage}
