+++ calliope/test/common/test_model/model_config_group/scenarios_energy_cap.yaml
@@ -1,0 +1,93 @@
+
+overrides:
+
+    energy_cap_min:
+        group_constraints.example_energy_cap_constraint:
+            energy_cap_min: 14
+
+    energy_cap_max:
+        group_constraints.example_energy_cap_constraint:
+            energy_cap_max: 6
+
+    energy_cap_loc_0:
+        group_constraints.example_energy_cap_constraint:
+            locs: ["0"]
+
+    energy_cap_min_expensive_supply:
+        group_constraints.example_energy_cap_min_constraint:
+            energy_cap_min: 4
+            techs: ["expensive_elec_supply"]
+
+    energy_cap_min_loc_0:
+        group_constraints.example_energy_cap_min_constraint:
+            locs: ["0"]
+
+    energy_cap_cheap_supply:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["cheap_elec_supply"]
+
+    energy_cap_expensive_supply:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["expensive_elec_supply"]
+
+    energy_cap_cheap_transmission:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["cheap_elec_transmission"]
+
+    energy_cap_expensive_transmission:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["expensive_elec_transmission"]
+
+    energy_cap_storage:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["elec_storage"]
+
+    energy_cap_elec_to_heat:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["elec_to_heat"]
+
+    energy_cap_elec_to_heat_cooling_linked:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["elec_to_heat_cool_linked"]
+
+    energy_cap_elec_to_heat_cooling_unlinked:
+        group_constraints.example_energy_cap_constraint:
+            techs: ["elec_to_heat_cool_unlinked"]
+
+scenarios:
+    energy_cap_max_supply: [energy_cap_max, energy_cap_cheap_supply]
+    energy_cap_min_supply:  [energy_cap_min, energy_cap_expensive_supply]
+    energy_cap_min_max_supply: [energy_cap_min_expensive_supply, energy_cap_max, energy_cap_cheap_supply]
+    energy_cap_max_supply_loc_0: [energy_cap_max, energy_cap_loc_0, energy_cap_cheap_supply]
+    energy_cap_min_supply_loc_0: [energy_cap_min_loc_0, energy_cap_min_expensive_supply]
+    energy_cap_min_max_supply_loc_0: [energy_cap_max, energy_cap_min_loc_0, energy_cap_min_expensive_supply, energy_cap_loc_0, energy_cap_cheap_supply]
+    energy_cap_max_transmission: [energy_cap_max, energy_cap_cheap_transmission]
+    energy_cap_min_transmission: [energy_cap_min, energy_cap_expensive_transmission, add_expensive_elec_transmission]
+    energy_cap_max_storage: [energy_cap_max, energy_cap_storage, add_elec_storage]
+    energy_cap_min_storage: [energy_cap_min, energy_cap_storage, add_elec_storage]
+    energy_cap_max_conversion: [energy_cap_max, energy_cap_elec_to_heat, add_cheap_heat_supply, add_elec_to_heat_conversion, add_heat_demand]
+    energy_cap_min_conversion: [energy_cap_min, energy_cap_elec_to_heat, add_cheap_heat_supply, add_elec_to_heat_conversion, add_heat_demand]
+    energy_cap_max_conversion_plus_linked: [
+        energy_cap_max, energy_cap_elec_to_heat_cooling_linked,
+        add_expensive_heat_supply, add_expensive_cooling_supply,
+        add_elec_to_heat_cooling_linked_conversion_plus,
+        add_heat_demand, add_cooling_demand
+    ]
+    energy_cap_min_conversion_plus_linked: [
+        energy_cap_min, energy_cap_elec_to_heat_cooling_linked,
+        add_cheap_heat_supply, add_cheap_cooling_supply,
+        add_elec_to_heat_cooling_linked_conversion_plus,
+        add_heat_demand, add_cooling_demand
+    ]
+    energy_cap_max_conversion_plus_unlinked: [
+        energy_cap_max, energy_cap_elec_to_heat_cooling_unlinked,
+        add_expensive_heat_supply, add_expensive_cooling_supply,
+        add_elec_to_heat_cooling_unlinked_conversion_plus,
+        add_heat_demand, add_cooling_demand
+    ]
+    energy_cap_min_conversion_plus_unlinked: [
+        energy_cap_min, energy_cap_elec_to_heat_cooling_unlinked,
+        add_cheap_heat_supply, add_cheap_cooling_supply,
+        add_elec_to_heat_cooling_unlinked_conversion_plus,
+        add_heat_demand, add_cooling_demand
+    ]
