+++ calliope/test/test_backend_pyomo.py
@@ -1835,60 +1835,6 @@ class TestPolicyConstraints:
         assert not hasattr(m._backend_model, 'group_share_carrier_prod_max_constraint')
         assert hasattr(m._backend_model, 'group_share_carrier_prod_equals_constraint')
 
-
-# Group constraints, i.e. those that can be defined on a system/subsystem scale
-@pytest.mark.filterwarnings("ignore:(?s).*Not all requested techs:calliope.exceptions.ModelWarning")
-class TestGroupConstraints:
-
-    def _build_group_model(self, scenario, model_file):
-        model = build_model(
-            model_file=model_file + '.yaml',
-            scenario=scenario
-        )
-        model.run(build_only=True)
-
-        return model
-
-    _test_vars = [
-        ('demand_share_max_systemwide', ['example_demand_share_max_constraint'], ['demand_share_max'], 'demand_share'),
-        ('demand_share_min_systemwide', ['example_demand_share_min_constraint'], ['demand_share_min'], 'demand_share'),
-        ('demand_share_max_location_0', ['example_demand_share_max_constraint'], ['demand_share_max'], 'demand_share'),
-        ('demand_share_min_location_0', ['example_demand_share_min_constraint'], ['demand_share_min'], 'demand_share'),
-        ('multiple_constraints', ['example_demand_share_max_constraint', 'example_demand_share_min_constraint'], ['demand_share_max', 'demand_share_min'], 'demand_share'),
-        ('cheap_cost_max_systemwide', ['example_cost_max_constraint'], ['cost_max'], 'model_cost_cap'),
-        ('expensive_cost_min_systemwide', ['example_cost_min_constraint'], ['cost_min'], 'model_cost_cap'),
-        ('cheap_cost_equals_systemwide', ['example_cost_equals_constraint'], ['cost_equals'], 'model_cost_cap'),
-        ('cheap_cost_max_location_0', ['example_cheap_cost_max_location_0_constraint'], ['cost_max'], 'model_cost_cap'),
-        ('multiple_costs_constraint', ['example_emissions_max_constraint', 'example_cost_min_constraint'], ['cost_max', 'cost_min'], 'model_cost_cap'),
-        ('cheap_cost_var_max_systemwide', ['example_cost_var_max_constraint'], ['cost_var_max'], 'model_cost_cap'),
-        ('cheap_cost_investment_max_systemwide', ['example_cost_investment_max_constraint'], ['cost_investment_max'], 'model_cost_cap')
-    ]
-
-    @pytest.mark.parametrize("scenario,group_name,constraints,model_file", _test_vars)
-    def test_constraint_names(self, scenario, group_name, constraints, model_file):
-        model = self._build_group_model(scenario, model_file)
-        assert all('group_names_' + i in model._model_data.dims for i in constraints)
-
-    @pytest.mark.parametrize("scenario,group_name,constraints,model_file", _test_vars)
-    def test_group_names(self, scenario, group_name, constraints, model_file):
-        model = self._build_group_model(scenario, model_file)
-        assert all(
-            group_name[i] in model._model_data['group_names_' + constraints[i]].values
-            for i in range(len(constraints))
-        )
-
-    @pytest.mark.parametrize("scenario,group_name,constraints,model_file", _test_vars)
-    def test_group_variables(self, scenario, group_name, constraints, model_file):
-        model = self._build_group_model(scenario, model_file)
-        assert all('group_' + i in model._model_data.data_vars.keys() for i in constraints)
-
-    @pytest.mark.parametrize("scenario,group_name,constraints,model_file", _test_vars)
-    def test_group_constraints(self, scenario, group_name, constraints, model_file):
-        model = self._build_group_model(scenario, model_file)
-        assert all(hasattr(model._backend_model, 'group_' + i + '_constraint')
-                   for i in constraints)
-
-
 # clustering constraints
 class TestClusteringConstraints:
